// src/types/styleInterface.ts
export interface CommonProps {
  $w?: string;
  $h?: string;
  $minW?: string;
  $maxW?: string;
  $minH?: string;
  $maxH?: string;

  $dis?: string;
  $gap?: string;
  $jus?: string;
  $mar?: string;
  $pad?: string;
  $flex?: string;
  $align?: string;
  $flexWrap?: string;

  $bor?: string;
  $borB?: string;
  $borT?: string;
  $borL?: string;
  $borR?: string;

  $over?: string;
  $overX?: string;
  $overY?: string;

  $wSpace?: string;
  $tTrans?: string;
  $direct?: string;
  $shadow?: string;
  $cursor?: string;
  $radius?: string;
  $tAlign?: string;
  $textShadow?: string;

  $aspect?: string;
  $objectFit?: string;

  $zIndex?: string;

  $top?: string;
  $left?: string;
  $right?: string;
  $trans?: string;
  $bottom?: string;
  $position?: string;

  $op?: string;
  $color?: string;
  $weight?: string;
  $textDeco?: string;
  $backColor?: string;
  $placeColor?: string;

  $size?: string;
  $line?: string;
  $font?: string;

  src: string;
  alt: any;

  $class?: string | string[];
}